Polyakov Alexander/telegram_bot/bot/requirements.txt
python-telegram-bot>=20.4
python-dotenv>=1
aiohttp>=3.8.5
PyYAML>=6.0.1